requirements.txt
aiohttp==3.7.4.post0
anyio==3.6.1
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cssselect==1.1.0
discord.py==1.7.3
fake-useragent==0.1.11
fastapi==0.78.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
multidict==6.0.2
parse==1.19.0
pydantic==1.9.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
requests==2.28.1
requests-html==0.10.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
w3lib==1.22.0
websockets==10.3
yarl==1.7.2
zipp==3.8.0